feat: disable nav items if namespace is not selected
