+++ src/AppRoot/AppSideBar.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Link, NavLink, useLocation, matchPath } from 'react-router-dom';
+import { Link, NavLink } from 'react-router-dom';
 import { Nav, NavItem, NavList, PageSidebar, PageSidebarBody } from '@patternfly/react-core';
 import { css } from '@patternfly/react-styles';
 import {
@@ -8,25 +8,10 @@ import {
   RELEASE_SERVICE_PATH,
   SECRET_LIST_PATH,
 } from '@routes/paths';
+import { useActiveRouteChecker } from '../../src/hooks/useActiveRouteChecker';
 import { useNamespace } from '../shared/providers/Namespace';
 
 import './AppSideBar.scss';
-
-export function useActiveRouteChecker() {
-  const location = useLocation();
-
-  /**
-   * Checks if the given route pattern matches the current location.
-   * @param pattern The route pattern to test, e.g. '/workspaces' or '/workspaces/:namespace/applications/*'
-   * @param options Options for matching. If exact is true, only an exact match is considered active.
-   * @returns True if the pattern is active, false otherwise.
-   */
-  return (pattern, options?: { exact?: boolean }): boolean => {
-    // Using React Router's matchPath to test the pattern against the current pathname.
-    const match = matchPath({ path: pattern, end: options?.exact ?? false }, location.pathname);
-    return Boolean(match);
-  };
-}
 
 export const AppSideBar: React.FC<{ isOpen: boolean }> = ({ isOpen }) => {
   const isActive = useActiveRouteChecker();
